Updated
git-svn-id: http://chianti.ucsd.edu/svn@23506 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ core3/table-import-impl/trunk/src/main/java/org/cytoscape/tableimport/internal/reader/AbstractGraphReader.java
@@ -38,7 +38,7 @@ public abstract class AbstractGraphReader implements GraphReader {
 	 *
 	 * @return  DOCUMENT ME!
 	 */
-	public int[] getNodeIndicesArray() {
+	public Long[] getNodeIndicesArray() {
 		return null;
 	}
 
@@ -47,7 +47,7 @@ public abstract class AbstractGraphReader implements GraphReader {
 	 *
 	 * @return  DOCUMENT ME!
 	 */
-	public int[] getEdgeIndicesArray() {
+	public Long[] getEdgeIndicesArray() {
 		return null;
 	}
